Splitted JS functionality, added browserify in a proper way
