+++ assets/css/style.css
@@ -109,6 +109,20 @@ section {
     list-style: none;
     margin: 0;
     padding-left: 0;
+    max-height: 3.5rem;
+    overflow: hidden;
+    cursor: pointer;
+    position: relative;
+}
+
+.streets::after {
+    position: absolute;
+    content: " ";
+    bottom: 0;
+    right: 0;
+    left: 0;
+    height: 1.8rem;
+    background-image: linear-gradient(-180deg, rgba(255, 255, 255, 0) 0%, rgba(255, 255, 255, 0.76) 100%);
 }
 
 .streets.toggled {
