+++ assets/js/main.js
@@ -1,0 +1,9 @@
+const toggle = require('./toggle')
+
+const app = {
+    init: function () {
+        toggle.init();
+    }
+};
+
+app.init();
