--- main.js
@@ -1,3 +1,0 @@
-require('leaflet')
-require('terraformer')
-require('terraformer-wkt-parser')
