+++ views/components/_layout.html
@@ -12,9 +12,6 @@
 {% block content %}
 {% endblock %}
 
-<script>
-    
-</script>
-
+<script src="js/bundle.js"></script>
 </body>
 </html>
